requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
blis==0.7.5
boto3==1.20.34
botocore==1.23.34
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.26
einops==0.3.0
filelock==3.4.2
fonttools==4.28.5
fsspec==2022.1.0
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.2.1
hydra-core==1.1.1
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
importlib-resources==5.4.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
murmurhash==1.0.6
networkx==2.6.3
nltk==3.6.7
numpy==1.21.4
oauthlib==3.1.1
olefile
omegaconf==2.1.1
opencv-python-headless==4.5.5.62
packaging==21.3
pathy==0.6.1
Pillow==9.0.0
preshed==3.0.6
progressbar2==3.55.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-utils==2.7.1
pytorch-lightning==1.0.8
pytorch-transformers==1.2.0
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.27.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
sacremoses==0.0.46
scikit-image==0.19.1
scipy==1.7.3
sentencepiece==0.1.96
six
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.0.13
tifffile==2021.11.2
tokenizers==0.10.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.15.0
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.7
warmup-scheduler==0.3
wasabi==0.9.0
Werkzeug==2.0.2
yacs==0.1.8
zipp==3.7.0
